Utilities/charge_slip_validator/pyproject.toml
[tool.poetry]
name = "charge-slip-validator"
version = "0.1.0"
description = ""
authors = ["Santo K Thomas"]
readme = "README.md"
packages = [{include = "charge_slip_validator"}]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.7.0.68"
jinja2 = ">=3.1.0"
numpy = "^1.24.1"
matplotlib = "^3.6.3"
pyyaml = "^6.0"
imutils = "^0.5.4"
selenium = "^4.8.0"
geckodriver-autoinstaller = "^0.1.0"
chromedriver-autoinstaller = "^0.4.0"
requests = "^2.28.2"
pytest-check = "^2.1.4"
sshtunnel = "^0.4.0"
pymysql = "^1.0.2"
pandas = "^1.5.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
